Reorganização das pastas e ajustando o fire do modal
