--- src/pages/members/assets/github.json
@@ -1,34 +1,0 @@
-{
-  "login": "VgsStudio",
-  "id": 81604963,
-  "node_id": "MDQ6VXNlcjgxNjA0OTYz",
-  "avatar_url": "https://avatars.githubusercontent.com/u/81604963?v=4",
-  "gravatar_id": "",
-  "url": "https://api.github.com/users/VgsStudio",
-  "html_url": "https://github.com/VgsStudio",
-  "followers_url": "https://api.github.com/users/VgsStudio/followers",
-  "following_url": "https://api.github.com/users/VgsStudio/following{/other_user}",
-  "gists_url": "https://api.github.com/users/VgsStudio/gists{/gist_id}",
-  "starred_url": "https://api.github.com/users/VgsStudio/starred{/owner}{/repo}",
-  "subscriptions_url": "https://api.github.com/users/VgsStudio/subscriptions",
-  "organizations_url": "https://api.github.com/users/VgsStudio/orgs",
-  "repos_url": "https://api.github.com/users/VgsStudio/repos",
-  "events_url": "https://api.github.com/users/VgsStudio/events{/privacy}",
-  "received_events_url": "https://api.github.com/users/VgsStudio/received_events",
-  "type": "User",
-  "site_admin": false,
-  "name": "Vitor Soller",
-  "company": null,
-  "blog": "",
-  "location": null,
-  "email": null,
-  "hireable": null,
-  "bio": "19 anos.\r\nEstudante de Engenharia da Computação no Instituto Mauá de Tecnologia. \r\nCurioso, dedicado, proativo e extrovertido nas horas vagas.",
-  "twitter_username": null,
-  "public_repos": 14,
-  "public_gists": 0,
-  "followers": 8,
-  "following": 11,
-  "created_at": "2021-03-30T03:39:05Z",
-  "updated_at": "2021-12-20T20:06:21Z"
-  }
