+++ src/pages/members/components/MembersContainer/MiniCardComponent.css
