+++ src/pages/members/components/MembersContainer/MiniCardComponent.jsx
@@ -6,18 +6,26 @@ import ModalTeste from "./ModalTeste";
 export default function MiniCardComponent(props) {
   const [modaCardlOpen, setModalCardOpen] = useState(false); // Abre o modal
 
-  function setModal() {
-    setModalCardOpen(!modaCardlOpen);
+  function setModal(e) {
+    console.log(e.target);
+    if (
+      e.target.className === "memberName" ||
+      e.target.className === "foto" ||
+      e.target.className === "members" ||
+      e.target.className === "modalBg"
+    ) {
+      setModalCardOpen(!modaCardlOpen);
+    }
   }
 
   return (
-    <div className="members" onClick={() => setModal()}>
+    <div className="members" onClickCapture={(e) => setModal(e)}>
       <h2 className="memberName">{props.memberInfo.name}</h2>
       <img src={props.memberInfo.avatar} alt="" className="foto" />
       <div className="socialContainer">
         <a
           className="social socialLinkedin"
-          href={props.member.linkedInLink}
+          href={props.member.linkedinLink}
           target="_blank"
           rel="noreferrer"
         >
@@ -25,7 +33,7 @@ export default function MiniCardComponent(props) {
         </a>
         <a
           className="social socialGithub"
-          href={props.member.linkedinLink}
+          href={props.member.githubLink}
           target="_blank"
           rel="noreferrer"
         >
