+++ src/pages/members/components/MembersContainer/ModalTeste.css
@@ -18,8 +18,6 @@
   height: 500px;
   background-color: red;
   z-index: 1001;
-
-  animation: ModalOpacity 200ms;
 }
 
 .modalBg {
@@ -37,4 +35,6 @@
 
 .modalCard {
   overflow: hidden;
+
+  animation: ModalOpacity 200ms;
 }
